+++ app/Http/Controllers/JobSectorController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\JobSector;
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Validation\ValidationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -44,10 +45,13 @@ class JobSectorController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function getList(Request $request)
     {
-        $filter =$this->jobSectorService->filterValidator($request)->validate();
+        $this->authorize('viewAny', JobSector::class);
+
+        $filter = $this->jobSectorService->filterValidator($request)->validate();
         try {
             $response = $this->jobSectorService->getJobSectorList($filter, $this->startTime);
         } catch (Throwable $e) {
@@ -65,6 +69,10 @@ class JobSectorController extends Controller
     {
         try {
             $response = $this->jobSectorService->getOneJobSector($id, $this->startTime);
+            if (!$response) {
+                abort(ResponseAlias::HTTP_NOT_FOUND);
+            }
+            $this->authorize('view', $response);
         } catch (Throwable $e) {
             return $e;
         }
@@ -76,9 +84,12 @@ class JobSectorController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     function store(Request $request): JsonResponse
     {
+        $this->authorize('create', JobSector::class);
+
         $validated = $this->jobSectorService->validator($request)->validate();
         try {
             $data = $this->jobSectorService->store($validated);
@@ -103,10 +114,14 @@ class JobSectorController extends Controller
      * @param int $id
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function update(Request $request, int $id): JsonResponse
     {
         $jobSector = JobSector::findOrFail($id);
+
+        $this->authorize('update', JobSector::class);
+
         $validated = $this->jobSectorService->validator($request, $id)->validate();
 
         try {
@@ -130,10 +145,13 @@ class JobSectorController extends Controller
      * Remove the specified resource from storage
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException
      */
     public function destroy(int $id): JsonResponse
     {
         $JobSector = JobSector::findOrFail($id);
+        $this->authorize('delete', $JobSector);
+
         try {
             $this->jobSectorService->destroy($JobSector);
             $response = [
@@ -196,7 +214,7 @@ class JobSectorController extends Controller
     {
         $JobSector = JobSector::onlyTrashed()->findOrFail($id);
         try {
-            $this->jobSectorService->forceDelete( $JobSector);
+            $this->jobSectorService->forceDelete($JobSector);
             $response = [
                 '_response_status' => [
                     "success" => true,
